server/requirements.txt
Flask==3.1.0
Flask-limiter==3.10.1
flasgger==0.9.7.1
gevent==23.9.1
gunicorn==20.1.0
pymongo==4.11.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
werkzeug==3.1.3
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Flask-Cors==5.0.0
